added AWS S3 images upload

+++ src/restaurants/restaurants.controller.ts
@@ -1,14 +1,28 @@
-import { Body, Controller, Delete, Get, Param, Post, Put, Query, UsePipes, ValidationPipe } from '@nestjs/common';
+import {
+  Body,
+  Controller,
+  Delete,
+  Get,
+  Param,
+  Post,
+  Put,
+  Query,
+  UploadedFiles,
+  UseInterceptors,
+  UsePipes,
+  ValidationPipe,
+} from '@nestjs/common';
 import { CreateRestaurantDto } from './dto/create-restaurant.dto';
 import { UpdateRestaurantDto } from './dto/update-restaurant.dto';
 import { RestaurantsService } from './restaurants.service';
 import { Restaurant } from './schemas/restaurant.schema';
 
-import {Query as ExpressQuery} from 'express-serve-static-core'
+import { Query as ExpressQuery } from 'express-serve-static-core';
+import { FilesInterceptor } from '@nestjs/platform-express';
 
 @Controller('restaurants')
 export class RestaurantsController {
-  constructor(private restaurantsService: RestaurantsService) { }
+  constructor(private restaurantsService: RestaurantsService) {}
 
   @Get()
   async getAllRestaurants(@Query() query: ExpressQuery): Promise<Restaurant[]> {
@@ -17,37 +31,67 @@ export class RestaurantsController {
 
   @Post()
   // @UsePipes(ValidationPipe) -> change to global
-  async createRestaurant(@Body() restaurant: CreateRestaurantDto): Promise<Restaurant>{
+  async createRestaurant(
+    @Body() restaurant: CreateRestaurantDto,
+  ): Promise<Restaurant> {
     return await this.restaurantsService.create(restaurant);
   }
 
   @Get(':id')
-  async getRestaurant(
-    @Param('id') id: string
-  ): Promise<Restaurant> {
+  async getRestaurant(@Param('id') id: string): Promise<Restaurant> {
     return this.restaurantsService.findById(id);
   }
 
   @Put(':id')
   async updateRestaurant(
     @Param('id') id: string,
-    @Body() restaurant: UpdateRestaurantDto
+    @Body() restaurant: UpdateRestaurantDto,
   ): Promise<Restaurant> {
-    await this.restaurantsService.findById(id)
-    return this.restaurantsService.updateById(id, restaurant)
+    await this.restaurantsService.findById(id);
+    return this.restaurantsService.updateById(id, restaurant);
   }
 
   @Delete(':id')
   async deleteRestaurant(
     @Param('id') id: string,
   ): Promise<{ deleted: Boolean }> {
-    await this.restaurantsService.findById(id)
-    const restaurant = await this.restaurantsService.deleteById(id)
-    console.log(restaurant)
-    if (restaurant) {
+    const restaurant = await this.restaurantsService.findById(id);
+    const isDeleted = await this.restaurantsService.deleteImages(
+      restaurant.images,
+    );
+    if (isDeleted) {
+      await this.restaurantsService.deleteById(id);
       return {
-        deleted: true
-      }
+        deleted: true,
+      };
+    } else {
+      return {
+        deleted: false,
+      };
     }
+  }
+
+  //To handle file uploading, Nest provides a built-in module based on the multer middleware package for Express.
+  @Put('upload/:id')
+  @UseInterceptors(FilesInterceptor('files'))
+  async uploadFiles(
+    @Param('id') id: string,
+    @UploadedFiles() files: Array<Express.Multer.File>,
+  ) {
+    await this.restaurantsService.findById(id);
+    return await this.restaurantsService.uploadImages(id, files);
+  }
+
+  //todo: deletId by images' key
+  @Put(':id/images/:imageKey')
+  async deleteFile(
+    @Param('id') id: string,
+    @Param('imageKey') imageKey: string,
+  ) {
+    const restaurant = await this.restaurantsService.findById(id);
+    const images = restaurant.images;
+
+    console.log(images);
+    images.filter((image) => {});
   }
 }
